Moved prisoners and basic code into separate files.
